Movement Script Update
I have update the PlayerMovement script to match the new simpler script. Also, I have changed one of the animations from isWalking to isRunning to match the object 'isRunning' in the script. Otherwise it conflicts and the object isRunning can not be detected in the animations. Or if y'all want we can do it vice versa, if we are going to have different animation for walking and running.